requirements.txt
tkinter
PyPDF2
pdf2image
pillow
requests
fpdf
python-dotenv
pymupdf